refactor: 핸들러 제거에 따른 패키징 수정
